requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
django-tinymce==4.0.0
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
